A lot of refactoring.
Moved the responsibility for reacting to input from drawable panes
to separate controllers.

Introduced a MenuStack, which is manipulated by NavigationMenuItems.
Not sure if it's a good idea to have NavigationMenuItems separate
from the rest -- seems kinda messy.

Packaged MainMenu and OptionsMenu into their own classes, with all the
creation code in their constructors, instead of in Application.
Application).

Packaged up a GamePane, too.